requirements.txt
google-genai>=0.8.0
langgraph>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.0